requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
PyQt5>=5.15.0
pandas>=1.3.0
openpyxl>=3.0.0
pyyaml>=5.4.0
scipy>=1.7.0
Pillow>=8.3.0
pyserial>=3.5
psutil>=5.8.0
pymodbus>=3.2.0
SQLAlchemy>=2.0.0
